add trace
